+++ connect-inject/container_init.go
@@ -8,17 +8,6 @@ import (
 	"text/template"
 
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-)
-
-// The init container is bound in memory usage by the size of the consul binary
-// as it issues a cpu of the binary to a shared volume. The limit is set to be
-// slightly larger than the binary to ensure we don't get OOM killed during the cp.
-const (
-	initContainerCPULimit      = "50m"
-	initContainerCPURequest    = "50m"
-	initContainerMemoryLimit   = "150Mi"
-	initContainerMemoryRequest = "25Mi"
 )
 
 type initContainerCommandData struct {
@@ -203,17 +192,6 @@ func (h *Handler) containerInit(pod *corev1.Pod, k8sNamespace string) (corev1.Co
 		return corev1.Container{}, err
 	}
 
-	resources := corev1.ResourceRequirements{
-		Limits: corev1.ResourceList{
-			corev1.ResourceCPU:    resource.MustParse(initContainerCPULimit),
-			corev1.ResourceMemory: resource.MustParse(initContainerMemoryLimit),
-		},
-		Requests: corev1.ResourceList{
-			corev1.ResourceCPU:    resource.MustParse(initContainerCPURequest),
-			corev1.ResourceMemory: resource.MustParse(initContainerMemoryRequest),
-		},
-	}
-
 	return corev1.Container{
 		Name:  "consul-connect-inject-init",
 		Image: h.ImageConsul,
@@ -251,7 +229,7 @@ func (h *Handler) containerInit(pod *corev1.Pod, k8sNamespace string) (corev1.Co
 				Value: fmt.Sprintf("$(POD_NAME)-%s", data.ProxyServiceName),
 			},
 		},
-		Resources:    resources,
+		Resources:    h.InitContainerResources,
 		VolumeMounts: volMounts,
 		Command:      []string{"/bin/sh", "-ec", buf.String()},
 	}, nil